bsp: simplify nrf53 specific implementation of pwm bsp library
