+++ bsp/bsp.emProject
@@ -79,7 +79,7 @@
       project_dependencies="00bsp_gpio"
       project_directory="."
       project_type="Library" />
-    <file file_name="nrf/$(PwmImplementationFile)" />
+    <file file_name="nrf/pwm.c" />
     <file file_name="pwm.h" />
   </project>
   <project Name="00bsp_qdec">
