+++ otap/dotbot-v2-bootloader.emProject
@@ -46,7 +46,7 @@
     linker_printf_fp_enabled="Float"
     linker_printf_width_precision_supported="Yes"
     linker_section_placement_file="$(ProjectDir)/../../nRF/Setup/nRF5340_xxAA_Application_flash_placement.xml"
-    macros="BuildTarget=dotbot-v2;PwmImplementationFile=pwm.c;DeviceFamily=nRF;Target=nRF5340_xxAA_Application"
+    macros="BuildTarget=dotbot-v2;DeviceFamily=nRF;Target=nRF5340_xxAA_Application"
     project_type="Executable"
     target_reset_script="Reset();"
     target_trace_initialize_script="EnableTrace(&quot;$(TraceInterfaceType)&quot;)" />
